升级Rust和依赖版本；更新API设计。

+++ src/server/user.rs
@@ -1,7 +1,7 @@
 use std::error::Error;
-use futures::{Stream, Future};
-use futures::future::lazy;
-use actix_web::{web, Scope, HttpRequest, HttpResponse, Error as ActixError};
+use futures::StreamExt;
+use actix_web::{web, Scope, HttpRequest, HttpResponse};
+use actix_multipart::Multipart;
 use super::super::model::user::{UpdateUser, UpdatePassword, ViewCover};
 use super::super::service::user::{user_get, user_update, user_set_password, user_set_cover};
 use super::super::service::image::Image;
@@ -57,38 +57,55 @@ fn set_password(body: web::Json<UpdatePassword>, req: HttpRequest) -> HttpRespon
     })
 }
 
-fn upload_cover(payload: web::Payload, req: HttpRequest) -> Box<dyn Future<Item=HttpResponse, Error=ActixError>> {
-    let user_id = match transaction_result(move|trans| verify_login(trans, &req)) { 
+async fn upload_cover(mut payload: Multipart, req: HttpRequest) -> HttpResponse {
+    let user_id = match transaction_result(move |trans| verify_login(trans, &req)) {
         Ok(user_id) => user_id,
-        Err(e) => return Box::new(lazy(||e)),
+        Err(e) => return e
     };
 
-    let image = match Image::new() { 
-        Ok(image) => image, 
-        Err(e) => return Box::new(lazy(move|| HttpResponse::InternalServerError().body(e.description().to_string())))
+    let mut image = match Image::new() {
+        Ok(image) => image,
+        Err(e) => return HttpResponse::InternalServerError().body(e.description().to_string())
     };
-    Box::new(payload.map_err(ActixError::from).fold(image, move|mut image, chunk| {
-        image.write(&chunk).unwrap();
-        Ok::<_, ActixError>(image)
-    }).and_then(move|mut image| {
-        let target = Image::new_filename(user_id);
-        let target_name = format!("{}/{}", get_config().get(STATIC_COVER_DIRECTORY), &target);
-        if let Err(e) = image.close().convert() { return HttpResponse::BadRequest().body(e.description().to_string()) }
-        if let Err(e) = image.move_to(&target_name) { return HttpResponse::InternalServerError().body(e.description().to_string()) }
-        image.clear();
-        transaction_res(|trans| {
-            if let Some(old_cover) = user_get(trans, user_id).unwrap().unwrap().cover {
-                if let Err(e) = Image::delete(&format!("{}/{}", get_config().get(STATIC_COVER_DIRECTORY), old_cover)) {
-                    return HttpResponse::InternalServerError().body(e.description().to_string())
-                }
+    let mut success = false;
+
+    while let Some(item) = payload.next().await {
+        let mut field = match item {
+            Ok(field) => field,
+            Err(e) => return HttpResponse::InternalServerError().body(e.to_string())
+        };
+        let disposition = field.content_disposition().unwrap();
+        let name = disposition.get_name().unwrap();
+        if name == "file" {
+            while let Some(chunk) = field.next().await {
+                let data = chunk.unwrap();
+                image.write(&data).unwrap();
             }
-            if let Err(e) = user_set_cover(trans, user_id, &target) { 
-                HttpResponse::InternalServerError().body(e.description().to_string()) 
-            }else{
-                HttpResponse::Ok().json(ViewCover{cover: target.to_string()})
+            success = true;
+            break;
+        }
+    }
+    if !success {
+        return HttpResponse::BadRequest().body("Item 'file' is necessary");
+    }
+
+    let target = Image::new_filename(user_id);
+    let target_name = format!("{}/{}", get_config().get(STATIC_COVER_DIRECTORY), &target);
+    if let Err(e) = image.close().convert() { return HttpResponse::BadRequest().body(e.description().to_string()) }
+    if let Err(e) = image.move_to(&target_name) { return HttpResponse::InternalServerError().body(e.description().to_string()) }
+    image.clear();
+    transaction_res(|trans| {
+        if let Some(old_cover) = user_get(trans, user_id).unwrap().unwrap().cover {
+            if let Err(e) = Image::delete(&format!("{}/{}", get_config().get(STATIC_COVER_DIRECTORY), old_cover)) {
+                return HttpResponse::InternalServerError().body(e.description().to_string())
             }
-        })
-    }))
+        }
+        if let Err(e) = user_set_cover(trans, user_id, &target) {
+            HttpResponse::InternalServerError().body(e.description().to_string())
+        }else{
+            HttpResponse::Ok().json(ViewCover{cover: target.to_string()})
+        }
+    })
 }
 
 pub fn register_view(scope: Scope) -> Scope {
@@ -97,5 +114,5 @@ pub fn register_view(scope: Scope) -> Scope {
         .route("/user/", web::post().to(set))
         .route("/user/", web::put().to(set))
         .route("/user/password/", web::post().to(set_password))
-        .route("/user/cover/", web::post().to_async(upload_cover))
+        .route("/user/cover/", web::post().to(upload_cover))
 }
